+++ Physics/Physics/Primitives/Intersectable.h
@@ -5,6 +5,7 @@
 namespace GLEngine::Physics::Primitives {
 template <class Derived> struct T_Intersectable {
 public:
+	virtual ~T_Intersectable() = default;
 	[[nodiscard]] S_RayIntersection Intersect(const S_Ray& ray) const { return static_cast<const Derived*>(this)->IntersectImpl(ray); }
 };
 } // namespace GLEngine::Physics::Primitives